Improve Terms and Conditions to include Benefits.
Line
From: "Throughout this Agreement, the term “Service” means the service allowing you to book pubs or offer your pub and pub services to others through our website or apps (the “Platform”). The term Service does not include any services provided by third parties. The terms “Customer” and “Pub” refer to the party Booking or Listing a Pub, respectively."
To: "Throughout this Agreement, the term “Service” means the service allowing you to book pubs or offer your pub and pub services to others through our website or apps (the “Platform”). It also means the service allowing you to benefit from the usage of our website or apps (the “Platform”), including benefits such as the Birthday Gift, the Corporate Benefit and the Recommend Reward, but not limited to them. The term Service does not include any services provided by third parties. The terms “Customer” and “Pub” refer to the party Booking or Listing a Pub, respectively."

Line
Added to 1.1 Services: "You may also be permitted to make use of the Whichost Benefits, such as the <a href="https://www.whichost.com/benefits/birthday-gift">Birthday Gift</a>, the <a href="https://www.whichost.com/benefits/corporate-benefit">Corporate Benefit</a> and the <a href="https://www.whichost.com/benefits/recommend-gift">Recommend Reward</a>, and other benefits that Whichost may implement in the future. You agree that your acceptance of this Whichost Terms of Service is also an acceptance of the terms and conditions that govern each of Whichost Benefits."